refactor exception mapper parse expected exception lists
