+++ src/ethereum_clis/__init__.py
@@ -4,7 +4,7 @@ Library of Python wrappers for the different implementations of transition tools
 
 from .clis.besu import BesuTransitionTool
 from .clis.ethereumjs import EthereumJSTransitionTool
-from .clis.evmone import EvmOneTransitionTool
+from .clis.evmone import EvmoneExceptionMapper, EvmOneTransitionTool
 from .clis.execution_specs import ExecutionSpecsTransitionTool
 from .clis.geth import GethTransitionTool
 from .clis.nimbus import NimbusTransitionTool
@@ -20,6 +20,7 @@ __all__ = (
     "EvmOneTransitionTool",
     "ExecutionSpecsTransitionTool",
     "GethTransitionTool",
+    "EvmoneExceptionMapper",
     "NimbusTransitionTool",
     "Result",
     "TransitionTool",
