+++ src/ethereum_test_exceptions/exception_mapper.py
@@ -1,12 +1,13 @@
 """
 EEST Exception mapper
 """
+
 from abc import ABC, abstractmethod
 from dataclasses import dataclass
 
 from bidict import frozenbidict
 
-from .exceptions import ExceptionBase, TransactionException, UndefinedException
+from .exceptions import ExceptionBase, UndefinedException
 
 
 @dataclass
@@ -54,37 +55,3 @@ class ExceptionMapper(ABC):
             exception_string, UndefinedException.UNDEFINED_EXCEPTION
         )
         return exception
-
-    def check_transaction(self, tx_error_message: str | None, tx, tx_pos: int):
-        """Verify transaction exception"""
-        exception_info = f"TransactionException (pos={tx_pos}, nonce={tx.nonce})\n"
-
-        if tx.error and not tx_error_message:
-            raise Exception(f"{exception_info} Error: tx expected to fail succeeded")
-        elif not tx.error and tx_error_message:
-            raise Exception(f"{exception_info} Error: tx unexpectedly failed: {tx_error_message}")
-        # TODO check exception list case
-        elif isinstance(tx.error, TransactionException) and tx_error_message:
-            expected_error_message = self.exception_to_message(tx.error)
-            error_exception = self.message_to_exception(tx_error_message)
-
-            define_message_hint = (
-                f"No message defined for {tx.error}, please add it to {self.__class__.__name__}"
-                if expected_error_message == "Undefined"
-                else ""
-            )
-            define_exception_hint = (
-                f"No exception defined for error message got, "
-                f"please add it to {self.__class__.__name__}"
-                if error_exception == UndefinedException.UNDEFINED_EXCEPTION
-                else ""
-            )
-
-            if expected_error_message not in tx_error_message:
-                raise Exception(
-                    f"{exception_info}"
-                    f"Error: exception mismatch:\n want = '{expected_error_message}' ({tx.error}),"
-                    f"\n got  = '{tx_error_message}' ({error_exception})"
-                    f"\n {define_message_hint}"
-                    f"\n {define_exception_hint}"
-                )
