+++ src/ethereum_test_specs/base.py
@@ -33,24 +33,6 @@ class HashMismatchException(Exception):
         return f"{self.message}: Expected {self.expected_hash}, got {self.actual_hash}"
 
 
-def verify_transactions(
-    exception_mapper: ExceptionMapper, txs: List[Transaction], result: Result
-) -> List[int]:
-    """
-    Verify rejected transactions (if any) against the expected outcome.
-    Raises exception on unexpected rejections or unexpected successful txs.
-    """
-    rejected_txs: Dict[int, str] = {
-        rejected_tx.index: rejected_tx.error for rejected_tx in result.rejected_transactions
-    }
-
-    for i, tx in enumerate(txs):
-        error_message = rejected_txs[i] if i in rejected_txs else None
-        exception_mapper.check_transaction(error_message, tx, i)
-
-    return list(rejected_txs.keys())
-
-
 def verify_result(result: Result, env: Environment):
     """
     Verify that values in the t8n result match the expected values.
